+++ src/main/java/io/dutwrapper/dutwrapper/HttpClientWrapper.java
@@ -1,8 +1,11 @@
 package io.dutwrapper.dutwrapper;
 
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
 import java.net.UnknownHostException;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import javax.annotation.Nullable;
@@ -12,13 +15,8 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 
+@SuppressWarnings("CallToPrintStackTrace")
 public class HttpClientWrapper {
-    private static boolean _showDebugLog = false;
-
-    public static void showDebugLog(boolean enabled) {
-        _showDebugLog = enabled;
-    }
-
     public static class Header {
         private String name;
         private String value;
@@ -81,7 +79,7 @@ public class HttpClientWrapper {
             return content;
         }
 
-        public void setContent(String content) {
+        public void setContent(@Nullable String content) {
             this.content = content;
         }
 
@@ -89,7 +87,7 @@ public class HttpClientWrapper {
             return exception;
         }
 
-        public void setException(Exception exception) {
+        public void setException(@Nullable Exception exception) {
             this.exception = exception;
         }
 
@@ -97,7 +95,7 @@ public class HttpClientWrapper {
             return message;
         }
 
-        public void setMessage(String message) {
+        public void setMessage(@Nullable String message) {
             this.message = message;
         }
 
@@ -105,7 +103,7 @@ public class HttpClientWrapper {
             return sessionId;
         }
 
-        public void setSessionId(String sessionId) {
+        public void setSessionId(@Nullable String sessionId) {
             this.sessionId = sessionId;
         }
 
@@ -144,12 +142,12 @@ public class HttpClientWrapper {
             throw new Exception("Not found session id in cookie!");
             // https://stackoverflow.com/questions/4959859/why-is-unknownhostexception-not-caught-in-exception-java
         } catch (UnknownHostException uheEx) {
-            if (_showDebugLog) {
+            if (Variables.getShowDebugLogStatus()) {
                 uheEx.printStackTrace();
             }
             return null;
         } catch (Exception ex) {
-            if (_showDebugLog) {
+            if (Variables.getShowDebugLogStatus()) {
                 ex.printStackTrace();
             }
             return null;
@@ -195,19 +193,19 @@ public class HttpClientWrapper {
             response.close();
             return result;
         } catch (NullPointerException nullEx) {
-            if (_showDebugLog) {
+            if (Variables.getShowDebugLogStatus()) {
                 nullEx.printStackTrace();
             }
             return new Response(
                     null, null, nullEx, nullEx.getMessage(), null);
         } catch (UnknownHostException uheEx) {
-            if (_showDebugLog) { 
+            if (Variables.getShowDebugLogStatus()) {
                 uheEx.printStackTrace();
             }
             return new Response(
                     null, null, uheEx, uheEx.getMessage(), null);
         } catch (Exception ex) {
-            if (_showDebugLog) {
+            if (Variables.getShowDebugLogStatus()) {
                 ex.printStackTrace();
             }
             return new Response(
@@ -259,17 +257,32 @@ public class HttpClientWrapper {
             response.close();
             return result;
         } catch (UnknownHostException uheEx) {
-            if (_showDebugLog) {                
+            if (Variables.getShowDebugLogStatus()) {
                 uheEx.printStackTrace();
             }
             return new Response(
                     null, null, uheEx, uheEx.getMessage(), null);
         } catch (Exception ex) {
-            if (_showDebugLog) {
+            if (Variables.getShowDebugLogStatus()) {
                 ex.printStackTrace();
             }
             return new Response(
                     null, null, ex, ex.getMessage(), null);
         }
+    }
+
+    public static byte[] toUrlEncode(Map<String, String> map, String charsetName) throws UnsupportedEncodingException {
+        StringBuilder request = new StringBuilder();
+
+        for (String key : map.keySet()) {
+            request.append(String.format(
+                    "%s%s=%s",
+                    request.length() != 0 ? "&" : "",
+                    URLEncoder.encode(key, "UTF-8"),
+                    URLEncoder.encode(map.get(key), "UTF-8")
+            ));
+        }
+
+        return request.toString().getBytes(charsetName);
     }
 }