+++ src/test/java/io/dutwrapper/dutwrapper/UtilsTest.java
@@ -4,13 +4,11 @@ import org.junit.jupiter.api.Test;
 
 import com.google.gson.Gson;
 
-import io.dutwrapper.dutwrapper.model.utils.DutSchoolYearItem;
-
 public class UtilsTest {
     @Test
     void finalTest() throws Exception {
-        DutSchoolYearItem test = Utils.getCurrentSchoolWeek();
+        Utils.DutSchoolYearItem test = Utils.getCurrentSchoolWeek();
         System.out.println(new Gson().toJson(test));
-        System.out.println(test.toString());
+        System.out.println(test);
     }
 }